\name{Compare-methods}
\docType{methods}
\alias{Compare-methods}
\alias{Compare,onion,onion-method}
\alias{Compare,onion,numeric-method}
\alias{Compare,numeric,onion-method}
\alias{onion_compare}
\title{Methods for Function Compare in Package Oniondingnag}
\description{Methods for comparision (greater than, etc) of onions}
\note{Comparison does not make sense for onions.}
\examples{
# roct() > 0 # meaningless and returns an error
  }
\keyword{methods}
\keyword{math}
