\name{cumsum}
\alias{cumsum}
\alias{onion_cumsum}
\alias{onion_cumprod}
\title{Cumulative sums and products of onions}
\description{
Cumulative sums and products of onions
}
\usage{
onion_cumsum(x)
onion_cumprod(x)
}
\arguments{\item{x}{onion}}
\author{Robin K. S. Hankin}
\examples{
cumsum(roct(5))
}
\keyword{array}
