\name{Logic}
\alias{logic.onion}
\alias{Logic,onion-method}
\alias{Logic,ANY,onion-method}
\alias{Logic,onion,ANY-method}
\alias{Logic,onion,missing-method}
\title{Logical operations on brobs}
\description{
Logical operations on onions are not supported
}
\author{Robin K. S. Hankin}
\note{

  The S4 group generic \dQuote{Logic} appeared in R-2.4.0-patched.

  Carrying out logical operations in this group will report an error.
    Negation, \dQuote{\code{!}}, is not part of this group.

}
\examples{
3+4
}
\keyword{math}
