\name{valid}
\alias{valid}
\alias{validity}
\alias{
\title{Validity of onionic vectors}
\description{Validity methods for quaternions and octonions}

\usage{
valid_quaternion(object)
valid_octonion(object)
}
\arguments{
  \item{object}{A numeric matrix, to be tested for suitability for
    coerction to an onion}
}
\author{Robin K. S. Hankin}
\examples{
2+3
}

