\name{Compare-methods}
\docType{methods}
\alias{Compare-methods}
\alias{Compare,onion,onion-method}
\alias{Compare,onion,numeric-method}
\alias{Compare,numeric,onion-method}
\alias{onion_compare}
\title{Methods for Function Compare in Package Oniondingnag}
\description{
  Methods for comparision (greater than, etc) of onions
}
\note{
  Comparison does not make sense for onions.
  
}
\examples{
2+2
  }
\keyword{methods}
\keyword{math}
