\name{Logic}
\alias{logic.onion}
\alias{onion_logic}
\alias{Logic,onion-method}
\alias{Logic,ANY,onion-method}
\alias{Logic,onion,ANY-method}
\alias{Logic,onion,missing-method}
\title{Logical operations on onions}
\description{Logical operations on onions are not supported}
\usage{onion_logic(e1,e2)}
\arguments{\item{e1,e2}{onions}}
\author{Robin K. S. Hankin}
\value{none}
\note{
  Carrying out logical operations in this group will report an error.
    Negation, \dQuote{\code{!}}, is not part of this group.

}
\examples{
# roct() & roct()   # reports an error
}
\keyword{math}
