\name{sum}
\alias{sum}
\alias{min}
\alias{max}
\alias{prod}
\alias{range}
\alias{Summary,onion-method}
\alias{sum,onion-method}
\alias{prod,quaternion-method}
\alias{sum,octonion-method}
\title{Various summary statistics for onions}
\description{Various summary statistics for onions}
\usage{
onion_allsum(x)
\S4method{sum}{onion}(x)
\S4method{prod}{quaternion}(x)
\S4method{sum}{octonion}(x)
}

\arguments{
  \item{x,\dots}{Objects of class \code{onion}}
}
\details{
  For a \code{onion} object
}
\author{Robin K. S. Hankin}
\note{
no notes here
}
\examples{
3+5
}
\keyword{math}
